# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.UpdateRoleRequest:
    properties:
      role:
        example: premium
        type: string
    required:
    - role
    type: object
...
  models.AuditLogsResponse:
    properties:
      count:
        type: integer
      limit:
        type: integer
      logs:
        items:
          $ref: '#/definitions/models.AuditLogResponse'
        type: array
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        description: |-
          HTTP status code
          example: 400
        example: 400
        type: integer
      details:
        description: Field-level validation error details (for validation errors)
        items:
          $ref: '#/definitions/models.FieldError'
        type: array
      error:
        description: |-
          Error type/category
          example: Bad Request
        example: Bad Request
        type: string
      message:
        description: |-
          Detailed error message
          example: Invalid email format
        example: Invalid email format
        type: string
      request_id:
        description: |-
          Request ID for tracing and debugging
          example: 550e8400-e29b-41d4-a716-446655440000
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  models.ImpersonateRequest:
    properties:
      reason:
        description: Reason for impersonation (for audit logging)
        type: string
      user_id:
        description: User ID to impersonate
        type: integer
    required:
    - user_id
    type: object
  models.ImpersonateResponse:
    properties:
      original_user_id:
        description: Original admin user ID (for stopping impersonation)
        type: integer
      token:
        description: New JWT token with impersonation claims
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.UserResponse'
        description: The user being impersonated
    type: object
...
  models.SuccessResponse:
    properties:
      data:
        description: Response data (varies by endpoint)
      message:
        description: |-
          Success message
          example: User created successfully
        example: User created successfully
        type: string
      success:
        description: |-
          Operation success status
          example: true
        example: true
        type: boolean
    type: object
...
  models.UserResponse:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: integer
      is_active:
        type: boolean
      location:
        type: string
      name:
        type: string
      oauth_provider:
        type: string
      role:
        type: string
      social_links:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/your-org/react-golang-starter-kit/issues
  description: A comprehensive REST API for the React Go Starter Kit application built
    with Fiber, GORM, and PostgreSQL. This API provides secure user authentication,
    user management, and system health monitoring.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/your-org/react-golang-starter-kit
  title: React Go Starter Kit API
  version: 1.0.0
paths:
  /admin/users/{id}/role:
    put:
      consumes:
      - application/json
      description: Update the role of a specific user by their ID. Requires admin
        or super_admin privileges.
      parameters:
      - description: User ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: New role for the user
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User role updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserResponse'
              type: object
        "400":
          description: Invalid user ID or role
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Failed to update user role
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a user's role (Admin only)
      tags:
      - admin
      - users
...
  /api/admin/audit-logs:
    get:
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Filter by target type
        in: query
        name: target_type
        type: string
      - description: Filter by action
        in: query
        name: action
        type: string
      - description: Filter by start date (RFC3339)
        in: query
        name: start_date
        type: string
      - description: Filter by end date (RFC3339)
        in: query
        name: end_date
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuditLogsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get audit logs
      tags:
      - Admin
...
  /api/admin/impersonate:
    post:
      parameters:
      - description: Impersonation request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ImpersonateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ImpersonateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start impersonating a user
      tags:
      - Admin
...
  /api/admin/stats:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AdminStatsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get admin statistics
      tags:
      - Admin
  /api/admin/stop-impersonate:
    post:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Stop impersonating
      tags:
      - Admin
  /api/admin/users:
    get:
      parameters:
      - description: Search query (name or email)
        in: query
        name: query
        required: true
        type: string
      - default: 10
        description: Max results
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search users
      tags:
      - Admin
  /api/admin/users/{id}/deactivate:
    post:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deactivate user
      tags:
      - Admin
  /api/admin/users/{id}/reactivate:
    post:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reactivate user
      tags:
      - Admin
  /api/admin/users/{id}/restore:
    post:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore deleted user
      tags:
      - Admin
  /api/admin/users/{id}/role:
    put:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: New role
        in: body
        name: body
        required: true
        schema:
          properties:
            role:
              type: string
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user role
      tags:
      - Admin
...
  /api/admin/users/deleted:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UsersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get deleted users
      tags:
      - Admin
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Format: `Authorization:
      Bearer {token}`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"